move precache workload cli into lca-cli
